# worklenz-backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Worklenz API
  version: 3.0.0
  description: |
    Worklenz project management platform API documentation.

    ## Authentication
    This API uses **session-based authentication** with **CSRF token protection**:

    1. **Session Cookie**: Automatically set after login
    2. **CSRF Token**: Required for all state-changing operations (POST, PUT, DELETE)
       - Obtain token from `GET /csrf-token` endpoint
       - Include in `X-CSRF-Token` header for subsequent requests

    ## Authorization Flow
    ```
    1. POST /auth/login → Receive session cookie
    2. GET /csrf-token → Receive CSRF token
    3. Use session + CSRF token for protected endpoints
    ```

    ## Response Format
    All endpoints return a standardized `ServerResponse<T>` object:
    ```json
    {
      "done": true,
      "body": { ... },
      "title": "Optional title",
      "message": "Optional message"
    }
    ```

  contact:
    name: Worklenz Support
    url: https://worklenz.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://app.worklenz.com
    description: Production server
...
security:
  - sessionAuth: []
    csrfToken: []
paths:
...
  /tasks/project/{id}:
    get:
      operationId: getTasksProjectById
      tags:
        - Tasks
      summary: Get tasks by project
      description: Retrieve all tasks for a specific project
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ServerResponse'
                  - type: object
                    properties:
                      body:
                        type: array
                        items:
                          $ref: '#/components/schemas/Task'
...
  /projects:
    get:
      operationId: getProjects
      tags:
        - Projects
      summary: List all projects
      description: Retrieve a list of all accessible projects with optional filtering
      security:
        - sessionAuth: []
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Pagination offset
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
          description: Number of projects to return
        - name: search
          in: query
          schema:
            type: string
          description: Search query to filter projects by name
      responses:
        '200':
          description: Projects retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ServerResponse'
                  - type: object
                    properties:
                      body:
                        type: array
                        items:
                          $ref: '#/components/schemas/Project'

    post:
      operationId: postProjects
      tags:
        - Projects
      summary: Create a new project
      description: Create a new project (requires team owner or admin role)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateRequest'
      responses:
        '200':
          description: Project created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ServerResponse'
                  - type: object
                    properties:
                      body:
                        $ref: '#/components/schemas/Project'
        '403':
          description: Insufficient permissions (requires team owner or admin)
  /projects/{id}:
    get:
      operationId: getProjectsById
      tags:
        - Projects
      summary: Get project by ID
      description: Retrieve detailed information about a specific project
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Project details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ServerResponse'
                  - type: object
                    properties:
                      body:
                        $ref: '#/components/schemas/Project'
        '404':
          description: Project not found

    put:
      operationId: putProjectsById
      tags:
        - Projects
      summary: Update project
      description: Update an existing project's properties (requires project manager role)
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdateRequest'
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ServerResponse'
                  - type: object
                    properties:
                      body:
                        $ref: '#/components/schemas/Project'
        '403':
          description: Insufficient permissions (requires project manager role)

    delete:
      operationId: deleteProjectsById
      tags:
        - Projects
      summary: Delete project
      description: Permanently delete a project (requires team owner or admin role)
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Project deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
              example:
                done: true
                body: null
                message: "Project deleted successfully"
        '403':
          description: Insufficient permissions
...
  /team-members/project/{id}:
    get:
      operationId: getTeamMembersProjectById
      tags:
        - Team Members
      summary: List members by project
      description: "List members by project. Includes request parameters, authorization expectations, and response details."
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Project members retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
  /team-members/projects/{id}:
    get:
      operationId: getTeamMembersProjectsById
      tags:
        - Team Members
      summary: List projects by team member
      description: "List projects by team member. Includes request parameters, authorization expectations, and response details."
      parameters:
        - $ref: '#/components/parameters/idParam'
        - name: project
          in: query
          required: false
          schema:
            type: string
          description: Comma-separated project IDs filter
        - name: status
          in: query
          required: false
          schema:
            type: string
          description: Comma-separated status IDs filter
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Member projects retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
...
  /reporting/overview/projects:
    get:
      tags:
        - Reporting
      operationId: getReportingOverviewProjects
      summary: Reporting overview projects
      description: "Reporting overview projects. Includes request parameters, authorization expectations, and response details."
      parameters:
        - name: team
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/reportingArchived'
        - $ref: '#/components/parameters/searchParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
      responses:
        '200':
          description: Projects retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
  /reporting/overview/projects/{team_id}:
    get:
      tags:
        - Reporting
      operationId: getReportingOverviewProjectsByTeam
      summary: Overview projects by team
      description: "Overview projects by team. Includes request parameters, authorization expectations, and response details."
      parameters:
        - $ref: '#/components/parameters/teamIdPathParam'
        - $ref: '#/components/parameters/memberQueryParam'
        - $ref: '#/components/parameters/reportingArchived'
      responses:
        '200':
          description: Projects retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
...
  /reporting/projects:
    get:
      tags:
        - Reporting
      operationId: getReportingProjects
      summary: Reporting projects
      description: Returns reporting project cards with rich filter support.
      parameters:
        - $ref: '#/components/parameters/reportingSearch'
        - $ref: '#/components/parameters/reportingStatuses'
        - $ref: '#/components/parameters/reportingHealths'
        - $ref: '#/components/parameters/reportingCategories'
        - $ref: '#/components/parameters/reportingProjectManagers'
        - $ref: '#/components/parameters/reportingTeams'
        - $ref: '#/components/parameters/reportingArchived'
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/pageSizeParam'
      responses:
        '200':
          description: Reporting projects retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
...
  /admin-center/organization/projects:
    get:
      tags:
        - Admin Center
      operationId: getOrganizationProjectsAdmin
      summary: List organization projects
      description: "List organization projects. Includes request parameters, authorization expectations, and response details."
      parameters:
        - $ref: '#/components/parameters/searchParam'
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/pageSizeParam'
      responses:
        '200':
          description: Projects retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie set after successful login

    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        CSRF token obtained from /csrf-token endpoint.
        Required for all POST, PUT, DELETE requests.

    clientPortalToken:
      type: apiKey
      in: header
      name: x-client-token
      description: Client portal authentication token (for client-specific endpoints)
  parameters:
    idParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Resource UUID
      example: "550e8400-e29b-41d4-a716-446655440000"
...
    searchParam:
      name: search
      in: query
      required: false
      schema:
        type: string
      description: Search query
...
    reportingArchived:
      name: archived
      in: query
      required: false
      schema:
        type: string
        enum: ["true", "false"]
      description: Include archived projects/tasks
  schemas:
    ServerResponse:
      type: object
      required:
        - done
        - body
      properties:
        done:
          type: boolean
          description: Indicates if the operation was successful
          example: true
        body:
          description: Response data (type varies by endpoint, null if no data)
          nullable: true
        title:
          type: string
          nullable: true
          description: Optional response title
        message:
          type: string
          nullable: true
          description: Optional message (error message or success notification)
          example: "Operation completed successfully"
...